Update: points to double att BD

+++ Database/Migrations/2022_01_25_097234632466_create_iauctions_bids_table.php
@@ -24,7 +24,7 @@ class CreateIauctionsBidsTable extends Migration
 
             $table->text('description');
             $table->double('amount', 30, 2)->default(0);
-            $table->float('points', 8, 2)->default(0);
+            $table->double('points', 30, 2)->default(0);
             $table->integer('status')->default(1)->unsigned();
 
             $table->boolean('winner')->default(false);
